refactor(mcp): move McpSessionStore into rb-mcp crate
Relocate the session store to crates/rb-mcp/src/session.rs so the MCP
protocol library owns its own runtime state. control-api/state.rs now
re-exports rb_mcp::McpSessionStore instead of defining it locally.

rb-mcp/Cargo.toml: add dashmap, uuid v4-feature; drop unused thiserror/chrono.

10 rb-mcp tests pass (4 new session tests); 280 control-api tests unchanged.
