+++ services/control-api/src/state.rs
@@ -3,10 +3,8 @@ use std::sync::{
     atomic::{AtomicI64, AtomicU64},
 };
 
-use dashmap::DashMap;
-use uuid::Uuid;
-
 use rb_auth::{LoginRateLimiter, PasswordHasher};
+pub use rb_mcp::McpSessionStore;
 use rb_email::EmailSender;
 use rb_github::GhApp;
 use rb_kafka::Producer;
@@ -18,36 +16,6 @@ use rb_storage_qdrant::TenantVectorStore;
 use sqlx::PgPool;
 
 use crate::config::Config;
-
-// ---------------------------------------------------------------------------
-// MCP session store (ADR-009 §1 — in-process, Phase 1)
-// ---------------------------------------------------------------------------
-
-/// In-memory table of active MCP sessions.
-///
-/// Keyed by the opaque session UUID returned in `Mcp-Session-Id`.  The value
-/// is the `tenant_id` that was bound at `initialize` time and is IMMUTABLE.
-/// Sessions are never evicted in Phase 1 (Phase 2 adds idle-timeout reaping).
-#[derive(Clone, Default)]
-pub struct McpSessionStore(Arc<DashMap<Uuid, Uuid>>);
-
-impl McpSessionStore {
-    pub fn new() -> Self {
-        Self(Arc::new(DashMap::new()))
-    }
-
-    /// Create a new session bound to `tenant_id` and return its ID.
-    pub fn create(&self, tenant_id: Uuid) -> Uuid {
-        let session_id = Uuid::new_v4();
-        self.0.insert(session_id, tenant_id);
-        session_id
-    }
-
-    /// Return the `tenant_id` for `session_id`, or `None` if not found.
-    pub fn tenant_id(&self, session_id: &Uuid) -> Option<Uuid> {
-        self.0.get(session_id).map(|r| *r)
-    }
-}
 
 // ---------------------------------------------------------------------------
 // Kafka consistency state
